+++ src/Repository/AbstractReadRepository.php
@@ -9,7 +9,6 @@ use Ramsey\Uuid\UuidInterface;
 use Sirix\Cycle\Extension\Domain\Contract\EntityInterface;
 use Sirix\Cycle\Extension\Domain\Contract\ReadRepositoryInterface;
 use Sirix\Cycle\Extension\Exception\EntityNotFoundException;
-use Sirix\Cycle\Extension\Factory\SelectFactory;
 
 use function is_int;
 
@@ -23,14 +22,6 @@ use function is_int;
 abstract class AbstractReadRepository extends Repository implements ReadRepositoryInterface
 {
     protected const USE_BINARY_UUID = true;
-
-    /**
-     * @param SelectFactory<TEntity> $selectFactory
-     */
-    public function __construct(SelectFactory $selectFactory)
-    {
-        parent::__construct($selectFactory($this->getEntityClass()));
-    }
 
     public function findByIdentifier(int|UuidInterface $identifier): ?EntityInterface
     {
@@ -47,11 +38,6 @@ abstract class AbstractReadRepository extends Repository implements ReadReposito
 
         return $entity;
     }
-
-    /**
-     * @return class-string<TEntity>
-     */
-    abstract protected function getEntityClass(): string;
 
     private function normalizePrimaryKey(int|UuidInterface $identifier): int|string
     {