+++ src/Repository/AbstractWriteRepository.php
@@ -6,10 +6,10 @@ namespace Sirix\Cycle\Extension\Repository;
 
 use Cycle\ORM\EntityManager;
 use Cycle\ORM\ORMInterface;
+use Cycle\ORM\Select;
 use Sirix\Cycle\Extension\Domain\Contract\EntityInterface;
 use Sirix\Cycle\Extension\Domain\Contract\WriteRepositoryInterface;
 use Sirix\Cycle\Extension\Exception\RepositoryPersistException;
-use Sirix\Cycle\Extension\Factory\SelectFactory;
 use Throwable;
 
 /**
@@ -24,11 +24,11 @@ abstract class AbstractWriteRepository extends AbstractReadRepository implements
     private readonly EntityManager $entityManager;
 
     /**
-     * @param SelectFactory<TEntity> $selectFactory
+     * @param Select<TEntity> $select
      */
-    public function __construct(protected readonly ORMInterface $orm, SelectFactory $selectFactory)
+    public function __construct(Select $select, protected readonly ORMInterface $orm)
     {
-        parent::__construct($selectFactory);
+        parent::__construct($select);
 
         $this->entityManager = new EntityManager($orm);
     }